+++ ks45team03/src/main/java/ks45team03/rentravel/dto/WaybillRenter.java
@@ -1,19 +1,20 @@
 package ks45team03.rentravel.dto;
 
 public class WaybillRenter {
-	private String waybillOwnerCode;
-	private String ownerId;
+	private String waybillRenterCode;
 	private String renterId;
+	private String ownerId;
 	private String paymentCode;
 	private String courierName;
 	private String waybillNumber;
 	private String waybillRegDate;
 	
-	public String getWaybillOwnerCode() {
-		return waybillOwnerCode;
+	
+	public String getWaybillRenterCode() {
+		return waybillRenterCode;
 	}
-	public void setWaybillOwnerCode(String waybillOwnerCode) {
-		this.waybillOwnerCode = waybillOwnerCode;
+	public void setWaybillRenterCode(String waybillRenterCode) {
+		this.waybillRenterCode = waybillRenterCode;
 	}
 	public String getOwnerId() {
 		return ownerId;
