+++ ks45team03/src/main/resources/mapper/infoBoardMapper.xml
@@ -1,0 +1,102 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="ks45team03.rentravel.mapper.InfoBoardMapper">
+	<resultMap type="InfoBoard" id="infoBoardResultMap">
+		<id property="infoBoardCode" 	  		column="info_board_code" />
+		<result property="userId" 	  			column="user_id" />
+		<result property="infoBoardTitle" 	  	column="info_board_title" />
+		<result property="infoBoardContent" 	column="info_board_content" />
+		<result property="infoBoardLikeCnt" 	column="info_board_like_cnt" />
+		<result property="infoBoardViewCnt" 	column="info_board_view_cnt" />
+		<result property="infoBoardRegTime" 	column="info_board_reg_time" />
+		<association property="user" javaType="User">
+			<id 	property="userId" 	  		column="user_id" />
+			<result property="userNickName" 	column="user_nickname" />
+			<result property="userImage" 	  	column="user_image" />
+		</association>
+	</resultMap>
+		
+	<resultMap type="InfoBoardComment" id="infoBoardCommentMap">
+		<id property="infoBoardCommentCode" 	  	column="info_board_comment_code" />
+		<result property="infoBoardCode" 	  		column="info_board_code" />
+		<result property="userId" 	  				column="user_id" />
+		<result property="userNickName" 			column="user_nickname" />
+		<result property="infoBoardCommentContent" 	column="info_board_comment_content" />
+		<result property="infoBoardCommentRegTime" 	column="info_board_comment_reg_time" />
+	</resultMap>
+	<!-- 정보 게시판 리스트 -->
+	<select id="getInfoBoardList" parameterType="map" resultMap="infoBoardResultMap">
+		SELECT
+			info.info_board_code,
+			info.info_board_title,
+			u.user_nickname,
+			info.info_board_like_cnt,
+			info.info_board_view_cnt,
+			info.info_board_reg_time
+		FROM
+			tb_info_board AS info
+			INNER JOIN
+			tb_user AS u
+			ON
+			info.user_id = u.user_id
+		ORDER BY
+			info_board_code DESC
+		<if test="startRowNum != null and startRowNum > -1">
+			LIMIT #{startRowNum}, #{rowPerPage};
+		</if>
+	</select>
+	<!-- 정보 게시판 리스트 총 개수-->
+	<select id="getInfoBoardListCnt" resultType="int">
+		SELECT
+			COUNT(1)
+		FROM
+			tb_info_board;
+	</select>
+	
+	<select id="getInfoBoardDetail" parameterType="String" resultMap="infoBoardResultMap">
+		SELECT
+			u.user_nickname,
+			info.info_board_code,
+			info.info_board_title,
+			info.info_board_content,
+			info.info_board_like_cnt,
+			info.info_board_view_cnt,
+			info.info_board_reg_time,
+			img.info_img
+		FROM
+			tb_info_board AS info
+			INNER JOIN
+			tb_info_board_img AS img
+			ON
+			img.info_board_code = info.info_board_code
+			INNER JOIN
+			tb_user AS u
+			ON
+			info.user_id = u.user_id
+		WHERE
+			info.info_board_code = #{infoBoardCode};
+	</select>
+	
+	<select id="getInfoBoardComment" parameterType="String" resultMap="infoBoardCommentMap">
+		SELECT
+			u.user_nickname,
+			comm.info_board_comment_code,
+			comm.user_id,
+			comm.info_board_comment_content,
+			comm.info_board_comment_reg_time
+		FROM
+			tb_info_board AS info
+			INNER JOIN
+			tb_info_board_comment AS comm
+			ON
+			info.info_board_code = comm.info_board_code
+			INNER JOIN
+			tb_user AS u
+			ON
+			comm.user_id = u.user_id
+		WHERE
+			comm.info_board_code = #{infoBoardCode};
+	</select>
+</mapper>
